+++ core/CHANGELOG.md
@@ -3,6 +3,7 @@
 - progress update callback on capture API options
 - API for querying supported devices
 - fix href parsing for URLs with query strings
+- API for parsing a page's anchor HREFs and checking HTTP statuses
 
 ## 0.0.3 - 2025-02-10